Update JSON Tutorial Code for Python 3
The existing code snippets which demonstrate the Transformer class rely
on tuple parameter unpacking, which was removed as of PEP-3113, which
landed in Python 3.0.

https://www.python.org/dev/peps/pep-3113/

The lark @v_args utility can be used at either the class or the
function level to provide behavior similar to tuple unpacking, which
improves the readability of the code.

The class methods don't make use of any internal data, so they can be
marked with @staticmethod.

The updated versions of these code snippets continue to work as expected
in Python 2.7.16, noting the difference in how unicode strings are
marked.

The code snippets copy and paste cleanly into the Python repl.