Add missing type hint for Token constructor
mypy thinks that Token's __init__ is inherited from __str__(). That's
not wrong -- it's just irrelevant, because Token also implements
__new__(). Token's _effective_ constructor signature is determined by
its __new__() method, so that's what I have used in the type hint.

Not clear if 'value' is supposed to be Any, but that's what I need in
my application. And it works just fine!