Improve documentation of `v_args` `meta` option.
When using `@v_args(meta=True)`, visitor methods will receive a `meta`
argument followed by a `children` argument in that order. The
documentation previously described the arguments in the opposite order,
which can be misleading to someone reading the docs, especially because
there was no example of `@v_args(meta=True)`.

This updates the documentation to describe the arguments in the order
they are passed in to the visitor methods, and adds a new example for
`@v_args(meta=True)`.